# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Solar Energy Project Documentation",
        "description": "Solar energy project management.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/install_type": {
            "get": {
                "tags": [
                    "InstallationType"
                ],
                "summary": "List install type with pagination.",
                "operationId": "acefeb6a4e1d6cf1d42d4f09d30af86f",
                "responses": {
                    "200": {
                        "description": "Install type successfully returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InstallationType"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Installation type found"
                    }
                }
            },
            "post": {
                "tags": [
                    "InstallationType"
                ],
                "summary": "Create a new Installation type",
                "operationId": "66c79092303b11e84fe3d26bb061ea36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item"
                                ],
                                "properties": {
                                    "item": {
                                        "type": "string",
                                        "example": "Laje"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New install type created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstallationType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/install_type/{id}": {
            "get": {
                "tags": [
                    "InstallationType"
                ],
                "summary": "Show a specific Installation type",
                "operationId": "8a74e1244c4b32e5f85ff1f22cfa31fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Installation type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Installation type successfully found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstallationType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Installation type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "InstallationType"
                ],
                "summary": "Delete a specific installation type",
                "operationId": "9afa2ad7716a8b2a4b1131607e20191a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete id install type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Install type successfully deleted"
                    },
                    "404": {
                        "description": "Install type not found"
                    }
                }
            }
        },
        "/api/install_type/{ID}": {
            "put": {
                "tags": [
                    "InstallationType"
                ],
                "summary": "Update a Installation type",
                "operationId": "48e69971cb73b4e424ba25f931c975c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Installation type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "item"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "3"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Campo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Installation type successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstallationType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InstallationType": {
                "required": [
                    "id",
                    "item"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "item": {
                        "type": "string",
                        "example": "Laje"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}